+++ supabase/migrations/20250608210344_add_rls_policy_for_tickets.sql
@@ -1,0 +1,8 @@
+CREATE POLICY "Allow authenticated users to insert their own tickets"
+ON public.tickets
+FOR INSERT
+WITH CHECK ((
+  SELECT auth.uid()
+) = (
+  SELECT supabase_id FROM public.users WHERE id = tickets.customer_id
+));